Move PID to /curtinfrc/
