+++ common/src/cpp/PID.cpp
@@ -1,4 +1,4 @@
-#include "PID.h"
+#include "curtinfrc/PID.h"
 
 PID::PID(double _kp, double _ki, double _kd, double _kde) {
   kp = _kp;
